Python/rsv/requirements.txt
azure-mgmt-recoveryservices
azure-identity
pyexcel-ods
azure-mgmt-recoveryservicesbackup